Rename `transitionMigrator` to `buttonIconTransitionMigrator`
