+++ src/blocks/button-maxi/button-maxi.js
@@ -29,8 +29,10 @@ import { buttonIcon } from '../../icons';
 /**
  * Migrators
  */
-import { blockMigrator } from '../../extensions/styles/migrators';
-import transitionMigrator from '../../extensions/styles/migrators/transitionMigrator';
+import {
+	blockMigrator,
+	buttonIconTransitionMigrator,
+} from '../../extensions/styles/migrators';
 
 /**
  * Block
@@ -61,6 +63,6 @@ registerBlockType('maxi-blocks/button-maxi', {
 		save,
 		prefix: 'button-',
 		selectors: customCss.selectors,
-		migrators: [transitionMigrator],
+		migrators: [buttonIconTransitionMigrator],
 	}),
 });
