+++ src/extensions/styles/migrators/index.js
@@ -1,3 +1,4 @@
 export { default as blockMigrator } from './blockMigrator';
 export { default as shapeDividerMigrator } from './shapeDividerMigrator';
-export { default as transitionMigrator } from './transitionMigrator';
+export { default as SVGTransitionMigrator } from './SVGTransitionMigrator';
+export { default as buttonIconTransitionMigrator } from './buttonIconTransitionMigrator';
